Also add KDC.Error
Dunno, maybe we want other types or something for the KDC module to
translate other errors into? The eventual goal is that all exceptions to
be caught be an exception handler get a class or so to listen on. Then
all other errors can just bubble up as uncaught exceptions.