Drop in a typed array polyfill
From http://www.calormen.com/polyfill/#typedarray

Issue #23. The way it does Uint8Array and friends is not the most
efficient, but these messages should be small and this basically only
effects IE <= 9 and pre-ICS Android.

If performance turns out to be a problem, we can do something more
interesting. (Possible solutions: monkey-patch in Uint8Array._getter and
_setter and only use those. Only use DataView. Use _setter where
appropriate and do something else for getter... have a
Uint8ArrayReadOnly or something that's secretly a plain array in the
polyfill case.)